add: 시간 포맷 함수 공통화
