+++ components/Comment.jsx
@@ -1,46 +1,10 @@
 import { useState } from "react";
+import { getRelativeTime } from "@/lib/dateUtils";
 import styles from "./Comment.module.css";
 
 function Comment({ comment, isMenuOpen, onMenuToggle, onModify, onDelete }) {
   const [isEditing, setIsEditing] = useState(false); // 수정 모드 여부
   const [editedContent, setEditedContent] = useState(comment.content); // 수정 내용
-
-  // 상대적 시간 계산 함수
-  const getRelativeTime = (createdAt) => {
-    const now = new Date();
-    const createdTime = new Date(createdAt);
-    const diffMs = now - createdTime;
-    const diffSeconds = Math.floor(diffMs / 1000);
-    const diffMinutes = Math.floor(diffSeconds / 60);
-    const diffHours = Math.floor(diffMinutes / 60);
-    const diffDays = Math.floor(diffHours / 24);
-
-    if (diffSeconds < 60) {
-      return `방금 전`;
-    } else if (diffMinutes < 20) {
-      return `10분 전`;
-    } else if (diffMinutes < 30) {
-      return `20분 전`;
-    } else if (diffMinutes < 40) {
-      return `30분 전`;
-    } else if (diffMinutes < 50) {
-      return `40분 전`;
-    } else if (diffMinutes < 60) {
-      return `50분 전`;
-    } else if (diffHours < 24) {
-      return `${diffHours}시간 전`;
-    } else if (diffDays < 4) {
-      return `${diffDays}일 전`;
-    } else {
-      return new Intl.DateTimeFormat("ko-KR", {
-        year: "2-digit",
-        month: "2-digit",
-        day: "numeric",
-      })
-        .format(createdTime)
-        .replace(/\.$/, "");
-    }
-  };
 
   const relativeTime = getRelativeTime(comment.createdAt);
   const isEdited = comment.updatedAt && comment.updatedAt !== comment.createdAt;
